requirements.txt
pandas
numpy
astral
soundfile
pyyaml
tqdm
tzdata
# birdnetlib (kommt später dazu)